requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic
python-jose
